Created the Bonus Ship
I made it where the bonus ship now appears and fires at the player, the ship will not yet take fire since we are implementing a new way of checking for collision. I implemented a new class called BonusShipManager with intent to add more into this file that the enemy manager would not be a suitable loaction for the data.