Added Power Ups
Added power ups that will be given randomly to the player. There are currently 4 that will be given along with an extra life. The 4 are an extra life(that makes two now), SpeedBoost, Sheild, and TripleBulletCap. These all will be given for an short amount of time(other than the extra life).